server/requirements.txt
altgraph==0.15
appdirs==1.4.3
asn1crypto==0.23.0
attrs==17.2.0
Automat==0.6.0
beautifulsoup4==4.6.0
bleach==2.0.0
bs4==0.0.1
CacheControl==0.12.5
cachetools==2.1.0
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
cheroot==6.0.0
CherryPy==14.0.0
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.1
cvxpy==1.0.9
cycler==0.10.0
decorator==4.1.2
dill==0.2.8.2
Django==2.0.1
ecos==2.0.5
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
firebase-admin==2.13.0
Flask==0.12.2
Flask-Cors==3.0.3
flask-marshmallow==0.9.0
Flask-SQLAlchemy==2.3.2
future==0.16.0
get==0.0.21
gevent==1.3.6
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.11.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
greenlet==0.4.15
html5lib==0.999999999
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipykernel==4.6.1
ipython==6.2.0
ipython-genutils==0.2.0
ipywidgets==7.0.1
itsdangerous==0.24
jdcal==1.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
lxml==4.0.0
macholib==1.9
MarkupSafe==1.0
marshmallow==2.15.6
marshmallow-sqlalchemy==0.14.1
matplotlib==2.0.2
mistune==0.7.4
more-itertools==4.1.0
msgpack==0.5.6
multiprocess==0.70.6.1
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.1.0
numpy==1.15.2
openpyxl==2.4.9
osqp==0.4.1
pandas==0.20.3
pandocfilters==1.4.2
parsel==1.2.0
parso==0.1.0
pefile==2017.11.5
pickleshare==0.7.4
ply==3.11
portend==2.2
post==0.0.13
prompt-toolkit==1.0.15
protobuf==3.6.1
public==0.0.38
PuLP==1.6.8
pyasn1==0.3.6
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyInstaller==3.3.1
PyJWT==1.6.4
pyodbc==4.0.18
Pyomo==5.4.3
pyOpenSSL==17.3.0
pyparsing==2.2.0
pypiwin32==220
pyreadline==2.1
python-dateutil==2.6.1
pytz==2017.2
PyUtilib==5.6.2
pywrap==0.1.0
pyzmq==16.0.2
qtconsole==4.3.1
query-string==0.0.12
queuelib==1.4.2
request==0.0.13
requests==2.18.4
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.0
service-identity==17.0.0
setupfiles==0.0.50
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.14
tempora==1.10
testpath==0.3.1
Theano==1.0.1
toolz==0.9.0
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
w3lib==1.18.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.3
xlrd==1.1.0
XlsxWriter==1.0.2
xlutils==2.0.0
xlwt==1.3.0
zope.interface==4.4.3
